Archive of features that got lost
This branch should not be deleted. It points to a long-diverged
collection of files that would be a nuisance to try to merge due
to file renames and code restructures, however, contains features
that may not be implemented in any more recent commits (smooth
stress recovery using Z2 error estimators, strain energy).Thus
this branch may be checked for useful code that may want copying
over if I ever remember already implementing something that has
been lost.